[#234] 로그인 안했을 때 alert 처리 구현
