+++ EATSSU/App/Sources/Presentation/CustomTabBarContainerController.swift
@@ -33,15 +33,21 @@ final class CustomTabBarContainerController: UIViewController {
         }
 
         tabBarView.buttonTapped = { [weak self] index in
-            self?.switchToViewController(at: index)
+            guard let self = self else { return }
+
+            // 로그인 요청 띄우기
+            if index == 2, RealmService.shared.isAccessTokenPresent() == false {
+                self.presentLoginAlert()
+                return
+            }
+
+            self.switchToViewController(at: index)
         }
     }
-
 
     private func switchToViewController(at index: Int) {
         let selectedVC = viewControllers[index]
 
-        // remove previous
         children.forEach { child in
             child.view.removeFromSuperview()
             child.removeFromParent()
@@ -58,4 +64,30 @@ final class CustomTabBarContainerController: UIViewController {
         tabBarView.setSelectedIndex(index)
         currentIndex = index
     }
+    
+    private func presentLoginAlert() {
+        let alert = UIAlertController(title: "로그인이 필요한 서비스입니다",
+                                      message: "로그인 하시겠습니까?",
+                                      preferredStyle: .alert)
+        let confirmAction = UIAlertAction(title: "확인", style: .default) { [weak self] _ in
+            self?.navigateToLogin()
+        }
+        let cancelAction = UIAlertAction(title: "취소", style: .cancel) { [weak self] _ in
+            guard let self = self else { return }
+            self.tabBarView.setSelectedIndex(self.currentIndex)
+        }
+        alert.addAction(confirmAction)
+        alert.addAction(cancelAction)
+        present(alert, animated: true, completion: nil)
+    }
+
+    private func navigateToLogin() {
+        let loginVC = LoginViewController()
+        if let windowScene = UIApplication.shared.connectedScenes.first as? UIWindowScene,
+           let sceneDelegate = windowScene.delegate as? SceneDelegate,
+           let window = sceneDelegate.window {
+            window.replaceRootViewController(loginVC)
+        }
+    }
+
 }
