+++ EATSSU/App/Sources/Presentation/Home/ViewController/HomeViewController.swift
@@ -42,7 +42,6 @@ final class HomeViewController: BaseViewController {
         configureUI()
         setLayout()
         registerTabman()
-//        setupNavigationBar()
     }
 
     override func viewDidAppear(_ animated: Bool) {
@@ -81,60 +80,6 @@ final class HomeViewController: BaseViewController {
             make.leading.trailing.bottom.equalToSuperview()
         }
         tabmanController.didMove(toParent: self)
-    }
-
-    // MARK: - Navigation
-
-//    private func setupNavigationBar() {
-//        let logoImageView = UIImageView(image: EATSSUDesignAsset.Images.mainLogoSmall.image)
-//        navigationItem.titleView = logoImageView
-//
-//        let rightButton = UIBarButtonItem(
-//            image: EATSSUDesignAsset.Images.myPageIcon.image,
-//            style: .plain,
-//            target: self,
-//            action: #selector(didTapRightBarButton)
-//        )
-//        rightButton.tintColor = EATSSUDesignAsset.Color.Main.primary.color
-//        navigationItem.rightBarButtonItem = rightButton
-//        navigationController?.isNavigationBarHidden = false
-//    }
-//
-//    @objc
-//    private func didTapRightBarButton() {
-//        if RealmService.shared.isAccessTokenPresent() {
-//            navigateToMyPage()
-//        } else {
-//            presentLoginAlert()
-//        }
-//    }
-
-    private func navigateToMyPage() {
-        let myPageVC = MyPageViewController()
-        navigationController?.pushViewController(myPageVC, animated: true)
-    }
-
-    private func presentLoginAlert() {
-        let alert = UIAlertController(title: "로그인이 필요한 서비스입니다",
-                                      message: "로그인 하시겠습니까?",
-                                      preferredStyle: .alert)
-        let confirmAction = UIAlertAction(title: "확인", style: .default) { [weak self] _ in
-            self?.navigateToLogin()
-        }
-        let cancelAction = UIAlertAction(title: "취소", style: .cancel, handler: nil)
-        alert.addAction(confirmAction)
-        alert.addAction(cancelAction)
-        present(alert, animated: true, completion: nil)
-    }
-
-    private func navigateToLogin() {
-        let loginVC = LoginViewController()
-        if let windowScene = UIApplication.shared.connectedScenes.first as? UIWindowScene,
-           let sceneDelegate = windowScene.delegate as? SceneDelegate,
-           let window = sceneDelegate.window
-        {
-            window.replaceRootViewController(loginVC)
-        }
     }
 
     // MARK: - Firebase
